web_basert_quiz/requirements.txt
Flask==2.1.3
click==8.0.4
colorama==0.4.4
dnspython==2.2.1
email-validator==1.1.3
Flask-Login==0.5.0
Flask-WTF==1.0.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
mysql-connector-python==8.0.28
protobuf==3.19.4
Werkzeug==2.0.3
WTForms==3.0.1
